+++ disconf-web/src/main/java/com/baidu/disconf/web/service/user/service/UserMgr.java
@@ -18,17 +18,17 @@ public interface UserMgr {
      * 
      * @return
      */
-    Visitor getVisitor(Integer userId);
+    Visitor getVisitor(Long userId);
 
     VisitorVo getCurVisitor();
 
-    User getUser(Integer userId);
+    User getUser(Long userId);
 
     /**
      * 
      * @return
      */
-    Integer create(User user);
+    Long create(User user);
 
     /**
      *
