+++ disconf-web/src/main/java/com/baidu/disconf/web/web/auth/login/impl/RedisLoginImplMock.java
@@ -28,7 +28,7 @@ public class RedisLoginImplMock implements RedisLogin {
     @Override
     public Visitor isLogin(HttpServletRequest request) {
 
-        int userId = 1;
+        Long userId = 1L;
 
         User user = userMgr.getUser(userId);
